add integrations test harness
The integrations crate provides a module test harness,
calling the substreams cli to run each of the modules
and cache the results to be compared on the next run.

Cached results are saved in the integrations/target/
folder and can be run locally. The cached results are
not committed to the repository, instead they can be
restored in the github ci workflow for comparison.

NOTE: There is some beginning work on calling the
substreams grpc blocks endpoint directly using the
substreams playground example as a guide, however, I
have not been able to call the endpoint successfully
and reverting to using the substreams CLI in the meantime.

Signed-off-by: Ryan Tate <ryan.michael.tate@gmail.com>